Remove unused code changes after merging the genrule commit
