+++ tensorboard/plugins/debugger_v2/BUILD
@@ -36,7 +36,6 @@ py_library(
         "//tensorboard:errors",
         "//tensorboard:plugin_util",
         "//tensorboard/backend:http_util",
-        "//tensorboard/compat/tensorflow_stub",
         "//tensorboard/plugins:base_plugin",
         "@org_pocoo_werkzeug",
     ],
