+++ tensorboard/plugins/debugger_v2/debugger_v2_plugin.py
@@ -22,7 +22,6 @@ from werkzeug import wrappers
 
 from tensorboard import errors
 from tensorboard import plugin_util
-from tensorboard.compat.tensorflow_stub import dtypes
 from tensorboard.plugins import base_plugin
 from tensorboard.plugins.debugger_v2 import debug_data_provider
 from tensorboard.backend import http_util
@@ -56,16 +55,6 @@ class DebuggerV2Plugin(base_plugin.TBPlugin):
         self._data_provider = debug_data_provider.LocalDebuggerV2DataProvider(
             self._logdir
         )
-        self._populate_dtypes_map()
-
-    def _populate_dtypes_map(self):
-        """Populate the map of dtypes, to prepare for later serving.
-
-        See `self.serve_dtypes_map()`.
-        """
-        self.dtypes_map = {}
-        for dtype_enum_value, dtype_string in dtypes._TYPE_TO_STRING.items():
-            self.dtypes_map[int(dtype_enum_value)] = {"name": dtype_string}
 
     def get_plugin_apps(self):
         # TODO(cais): Add routes as they are implemented.
@@ -77,7 +66,6 @@ class DebuggerV2Plugin(base_plugin.TBPlugin):
             "/source_files/list": self.serve_source_files_list,
             "/source_files/file": self.serve_source_file,
             "/stack_frames/stack_frames": self.serve_stack_frames,
-            "/dtypes_map": self.serve_dtypes_map,
         }
 
     def is_active(self):
@@ -289,17 +277,3 @@ class DebuggerV2Plugin(base_plugin.TBPlugin):
             )
         except errors.NotFoundError as e:
             return _error_response(request, str(e))
-
-    @wrappers.Request.application
-    def serve_dtypes_map(self, request):
-        """Serves the map from dtype num value to dtype details.
-
-        Args:
-          request: HTTP request.
-
-        Returns:
-          Response to the request.
-        """
-        return http_util.Respond(
-            request, {"dtypes_map": self.dtypes_map,}, "application/json",
-        )
