+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/actions/debugger_actions.ts
@@ -24,7 +24,6 @@ import {
   StackFramesById,
 } from '../store/debugger_types';
 import {
-  DtypesMapResponse,
   ExecutionDigestsResponse,
   ExecutionDataResponse,
 } from '../data_source/tfdbg2_data_source';
@@ -41,15 +40,6 @@ import {
  * Actions for the Debugger Component.
  */
 export const debuggerLoaded = createAction('[Debugger] Debugger Loaded');
-
-export const dtypesMapRequested = createAction(
-  '[Debugger] DTypes Map Requested'
-);
-
-export const dtypesMapLoaded = createAction(
-  '[Debugger] DTypes Map Loaded',
-  props<DtypesMapResponse>()
-);
 
 export const debuggerRunsRequested = createAction(
   '[Debugger] Debugger Runs Requested'
