+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/data_source/tfdbg2_data_source.ts
@@ -17,7 +17,6 @@ import {Observable} from 'rxjs';
 import {
   Alert,
   DebuggerRunListing,
-  DtypesMap,
   Execution,
   ExecutionDigest,
   StackFrame,
@@ -25,10 +24,6 @@ import {
 import {TBHttpClient} from '../../../../webapp/webapp_data_source/tb_http_client';
 
 /** @typehack */ import * as _typeHackRxjs from 'rxjs';
-
-export interface DtypesMapResponse {
-  dtypes_map: DtypesMap;
-}
 
 // The backend route for source-file list responds with an array
 // of 2-tuples: <host_name, file_path>.
@@ -75,8 +70,6 @@ export interface AlertsResponse {
 export abstract class Tfdbg2DataSource {
   abstract fetchRuns(): Observable<DebuggerRunListing>;
 
-  abstract fetchDtypesMap(): Observable<DtypesMapResponse>;
-
   abstract fetchExecutionDigests(
     run: string,
     begin: number,
@@ -122,12 +115,6 @@ export class Tfdbg2HttpServerDataSource implements Tfdbg2DataSource {
   private readonly httpPathPrefix = 'data/plugin/debugger-v2';
 
   constructor(private http: TBHttpClient) {}
-
-  fetchDtypesMap(): Observable<DtypesMapResponse> {
-    return this.http.get<DtypesMapResponse>(
-      this.httpPathPrefix + '/dtypes_map'
-    );
-  }
 
   fetchRuns() {
     // TODO(cais): Once the backend uses an DataProvider that unifies tfdbg and
