+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/effects/debugger_effects.ts
@@ -30,8 +30,6 @@ import {
   debuggerLoaded,
   debuggerRunsRequested,
   debuggerRunsLoaded,
-  dtypesMapLoaded,
-  dtypesMapRequested,
   executionDataLoaded,
   executionDigestFocused,
   executionDigestsRequested,
@@ -63,7 +61,6 @@ import {
   getLoadedExecutionData,
   getLoadedStackFrames,
   getNumAlertsOfFocusedType,
-  getDtypesMapLoaded,
   getSourceFileListLoaded,
 } from '../store/debugger_selectors';
 import {
@@ -77,7 +74,6 @@ import {
 } from '../store/debugger_types';
 import {
   AlertsResponse,
-  DtypesMapResponse,
   Tfdbg2HttpServerDataSource,
   SourceFileListResponse,
 } from '../data_source/tfdbg2_data_source';
@@ -177,28 +173,6 @@ export class DebuggerEffects {
             this.store.dispatch(
               debuggerRunsLoaded({runs: runs as DebuggerRunListing})
             );
-          }),
-          map(() => void null)
-          // TODO(cais): Add catchError() to pipe.
-        );
-      })
-    );
-  }
-
-  /**
-   * Load dtypes info when debugger plugin is loaded.
-   */
-  private loadDtypesMap(prevStream$: Observable<void>) {
-    return prevStream$.pipe(
-      withLatestFrom(this.store.select(getDtypesMapLoaded)),
-      filter(([, dtypesMapLoadState]) => {
-        return dtypesMapLoadState.state !== DataLoadState.LOADING;
-      }),
-      tap(() => this.store.dispatch(dtypesMapRequested())),
-      mergeMap(() => {
-        return this.dataSource.fetchDtypesMap().pipe(
-          tap((dtypesMapResponse: DtypesMapResponse) => {
-            this.store.dispatch(dtypesMapLoaded(dtypesMapResponse));
           }),
           map(() => void null)
           // TODO(cais): Add catchError() to pipe.
@@ -755,7 +729,6 @@ export class DebuggerEffects {
 
         // ExecutionDigest and ExecutionData can be loaded in parallel.
         return merge(
-          loadDtypesMap$,
           onNumAlertsLoaded$,
           onExcutionDigestLoaded$,
           onExecutionDataLoaded$,
