+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/store/debugger_selectors.ts
@@ -29,7 +29,6 @@ import {
   StackFrame,
   StackFramesById,
   State,
-  DtypesMap,
 } from './debugger_types';
 
 // HACK: These imports are for type inference.
@@ -56,20 +55,6 @@ export const getDebuggerRunsLoaded = createSelector(
 export const getActiveRunId = createSelector(
   selectDebuggerState,
   (state: DebuggerState): string | null => state.activeRunId
-);
-
-export const getDtypesMapLoaded = createSelector(
-  selectDebuggerState,
-  (state: DebuggerState): LoadState => {
-    return state.dtypes.dtypesLoaded;
-  }
-);
-
-export const getDtypesMap = createSelector(
-  selectDebuggerState,
-  (state: DebuggerState): DtypesMap => {
-    return state.dtypes.dtypesMap;
-  }
 );
 
 export const getAlertsLoaded = createSelector(
