+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/testing/index.ts
@@ -93,13 +93,6 @@ export function createDebuggerState(
       state: DataLoadState.NOT_LOADED,
       lastLoadedTimeInMs: null,
     },
-    dtypes: {
-      dtypesMap: {},
-      dtypesLoaded: {
-        state: DataLoadState.NOT_LOADED,
-        lastLoadedTimeInMs: null,
-      },
-    },
     activeRunId: null,
     alerts: createAlertsState(),
     executions: createDebuggerExecutionsState(),
